refactor: add comments & restructure project
