+++ src/commands/admin/give_mod.rs
@@ -8,8 +8,8 @@ use serenity::{prelude::*, utils::parse_mention};
 #[command]
 #[checks(Godmod)]
 #[only_in(guilds)]
-#[description = "Gives a user the mod role."]
 #[aliases("givemod")]
+/// Gives a user the mod role.
 pub async fn give_mod(ctx: &Context, msg: &Message, args: Args) -> CommandResult {
 	let mention = args.rest();
 
@@ -18,7 +18,10 @@ pub async fn give_mod(ctx: &Context, msg: &Message, args: Args) -> CommandResult
 		.await
 		.expect("Accessing the guild broke.");
 
+	// Get the id of the user mentioned
 	let id = parse_mention(mention).expect("Mention was invalid");
+
+	// Remove the mod role from the user
 	let mut member = guild.member(&ctx.http, id).await?;
 	member.add_role(&ctx.http, MOD_ROLE_ID).await?;
